requirements.txt
python-dotenv
groq
PyMuPDF
openai
pdfplumber
streamlit

acres==0.5.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
# cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
narwhals==1.46.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
openai==1.95.1
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
protobuf==6.31.1
prov==2.1.1
puremagic==1.30
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0


